+++ components/remote-controls/grid/remote/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/grid/remote/presentation/decompose/internal/RemoteGridScreenDecomposeComponentImpl.kt
@@ -5,6 +5,7 @@ import com.arkivanov.decompose.ComponentContext
 import com.arkivanov.decompose.childContext
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.keyedit.api.NotSavedFlipperKey
+import com.flipperdevices.remotecontrols.api.FlipperDispatchDialogApi
 import com.flipperdevices.remotecontrols.api.model.ServerRemoteControlParam
 import com.flipperdevices.remotecontrols.grid.remote.api.RemoteGridScreenDecomposeComponent
 import com.flipperdevices.remotecontrols.impl.grid.remote.composable.RemoteGridComposable
@@ -20,7 +21,8 @@ class RemoteGridScreenDecomposeComponentImpl @AssistedInject constructor(
     @Assisted param: ServerRemoteControlParam,
     @Assisted onBack: DecomposeOnBackParameter,
     @Assisted onSaveKey: (NotSavedFlipperKey) -> Unit,
-    remoteGridComponentFactory: RemoteGridComponent.Factory
+    remoteGridComponentFactory: RemoteGridComponent.Factory,
+    flipperDispatchDialogApiFactory: FlipperDispatchDialogApi.Factory,
 ) : RemoteGridScreenDecomposeComponent(componentContext) {
     private val gridComponent = remoteGridComponentFactory.invoke(
         componentContext = childContext("GridComponent"),
@@ -28,9 +30,13 @@ class RemoteGridScreenDecomposeComponentImpl @AssistedInject constructor(
         onBack = onBack,
         onSaveKey = onSaveKey
     )
+    private val flipperDispatchDialogApi = flipperDispatchDialogApiFactory.invoke(onBack = onBack)
 
     @Composable
     override fun Render() {
-        RemoteGridComposable(gridComponent)
+        RemoteGridComposable(
+            remoteGridComponent = gridComponent,
+            flipperDispatchDialogApi = flipperDispatchDialogApi
+        )
     }
 }