+++ components/remote-controls/grid/remote/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/grid/remote/presentation/mapping/GridComponentStateMapper.kt
@@ -2,6 +2,7 @@ package com.flipperdevices.remotecontrols.impl.grid.remote.presentation.mapping
 
 import com.flipperdevices.infrared.api.InfraredConnectionApi
 import com.flipperdevices.remotecontrols.api.DispatchSignalApi
+import com.flipperdevices.remotecontrols.api.FlipperDispatchDialogApi.Companion.toDialogType
 import com.flipperdevices.remotecontrols.api.SaveTempSignalApi
 import com.flipperdevices.remotecontrols.impl.grid.remote.presentation.decompose.RemoteGridComponent
 import com.flipperdevices.remotecontrols.impl.grid.remote.presentation.viewmodel.RemoteGridViewModel
@@ -23,7 +24,7 @@ internal object GridComponentStateMapper {
                     RemoteGridComponent.Model.Loaded(
                         pagesLayout = gridState.pagesLayout,
                         remotes = gridState.remotes,
-                        isFlipperBusy = dispatchState is DispatchSignalApi.State.FlipperIsBusy,
+                        flipperDialog = dispatchState.toDialogType(),
                         emulatedKey = (dispatchState as? DispatchSignalApi.State.Emulating)?.ifrKeyIdentifier,
                         saveState = saveState,
                         connectionState = connectionState