+++ components/remote-controls/setup/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/setup/composable/SetupScreen.kt
@@ -15,10 +15,10 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import com.flipperdevices.core.ui.dialog.composable.busy.ComposableFlipperBusy
 import com.flipperdevices.core.ui.theme.LocalPalletV2
 import com.flipperdevices.ifrmvp.core.ui.layout.shared.ErrorComposable
 import com.flipperdevices.ifrmvp.core.ui.layout.shared.SharedTopBar
+import com.flipperdevices.remotecontrols.api.FlipperDispatchDialogApi
 import com.flipperdevices.remotecontrols.impl.setup.composable.components.AnimatedConfirmContent
 import com.flipperdevices.remotecontrols.impl.setup.composable.components.LoadedContent
 import com.flipperdevices.remotecontrols.impl.setup.composable.components.SetupLoadingContent
@@ -40,6 +40,7 @@ private val SetupComponent.Model.key: Any
 @Composable
 fun SetupScreen(
     setupComponent: SetupComponent,
+    flipperDispatchDialogApi: FlipperDispatchDialogApi,
     modifier: Modifier = Modifier
 ) {
     val rootNavigation = LocalRootNavigation.current
@@ -81,15 +82,10 @@ fun SetupScreen(
                 }
 
                 is SetupComponent.Model.Loaded -> {
-                    if (model.isFlipperBusy) {
-                        ComposableFlipperBusy(
-                            onDismiss = setupComponent::dismissBusyDialog,
-                            goToRemote = {
-                                setupComponent.dismissBusyDialog()
-                                rootNavigation.push(RootScreenConfig.ScreenStreaming)
-                            }
-                        )
-                    }
+                    flipperDispatchDialogApi.Render(
+                        dialogType = model.flipperDialog,
+                        onDismiss = setupComponent::dismissDialog
+                    )
                     LoadedContent(
                         model = model,
                         modifier = Modifier.padding(scaffoldPaddings),