+++ CHANGELOG.md
@@ -10,8 +10,10 @@ Attention: don't forget to add the flag for F-Droid before release
 - [Feature] Add onetap widget
 - [Feature] Save, edit, share remote control
 - [Feature] More UI elements for remote-controls
+- [Feature] Add How to Use dialog into remote-controls
 - [Feature] Skip infrared signals on setup screen
-- [Feature] Add new icons for remote-controls
+- [Feature] Better user-ux when configuring remote control
+- [Feature] Add flipper action dialogs into remote control and move it into bottombar
 - [Refactor] Load RemoteControls from flipper, emulating animation
 - [Refactor] Update to Kotlin 2.0
 - [Refactor] Replace Ktorfit with Ktor requests in remote-controls